refactor: Move aqua package
