+++ pkg/aqua/client/auth.go
