+++ pkg/aqua/scanner/cli/types.go
