+++ pkg/aqua/scanner/testdata/aqua_csp_report/without_vulnerabilities.json
